Update the best thing 10.py

+++ the best thing 10.py
@@ -14,7 +14,7 @@ else:
 if 'level' not in st.session_state:
   st.session_state.level=file
 if 'ran' not in st.session_state or st.session_state.ran < 1:
-  st.session_state.ran=20*st.session_state.level
+  st.session_state.ran=int(20*st.session_state.level)
 if 'num' not in st.session_state:
   st.session_state.num=0
 if 'sc' not in st.session_state:
